trying to fix the POST methods

+++ microservices/second-service/src/main/java/com/miniprojet/secondservice/web/RendezvousController.java
@@ -1,9 +1,11 @@
 package com.miniprojet.secondservice.web;
 
 import com.miniprojet.secondservice.client.FirstServiceRestClient;
+import com.miniprojet.secondservice.entities.Consultation;
 import com.miniprojet.secondservice.entities.RendezVous;
 import com.miniprojet.secondservice.model.Medecin;
 import com.miniprojet.secondservice.model.Patient;
+import com.miniprojet.secondservice.repositories.ConsultationRepository;
 import com.miniprojet.secondservice.repositories.RendezvousRepository;
 import com.miniprojet.secondservice.services.RendezvousService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,8 @@ import java.util.List;
 public class RendezvousController {
     @Autowired
     private RendezvousService rendezvousService;
+    @Autowired
+    private ConsultationRepository consultationRepository;
 //    @Autowired
 //    private MedecinRestClient medecinRestClient;
     @Autowired
@@ -43,12 +47,19 @@ public class RendezvousController {
     //DELETE
     @DeleteMapping("rendezVous/{rendezVousId}")
     public void deleteRendezVous(@PathVariable  Long rendezVousId){
-        rendezvousRepository.deleteById(rendezVousId);
-//        rendezvousService.DeleteRendezVous(rendezVousId);
+        RendezVous rendezVous = rendezvousRepository.findById(rendezVousId).orElse(null);
+        if (rendezVous != null) {
+            Consultation consultation = rendezVous.getConsultation();
+            if (consultation != null) {
+                consultationRepository.deleteById(consultation.getId());
+            }
+            rendezvousRepository.deleteById(rendezVousId);
+        }
     }
     //CREATE
     @PostMapping("/rendezVous")
     public RendezVous saveRendezVous(@RequestBody RendezVous rendezVous){
+        consultationRepository.save(rendezVous.getConsultation());
         return rendezvousService.saveRendezVous(rendezVous);
     }
     //UPDATE
